+++ tst/list.js
@@ -29,7 +29,7 @@ ct('list-simple', function() {
 	ct('===', concatData(el), '^4312$')//2
 	comp.update([1,5,3])
 	ct('===', concatData(el), '^153$')//3
-	ct('===', l0.clear(), l0)
+	ct('===', l0.removeChildren(), l0)
 	ct('===', concatData(el), '^$')
 	ct('===', l0.moveTo(null), l0)
 	ct('===', concatData(el), '')